fix: authorize relation read endpoints
