+++ src/Laravel/src/Http/Requests/Relations/RelationModelFieldRequest.php
@@ -9,12 +9,15 @@ use Illuminate\Foundation\Http\FormRequest;
 use MoonShine\Contracts\Core\CrudPageContract;
 use MoonShine\Contracts\Core\CrudResourceContract;
 use MoonShine\Contracts\UI\HasFieldsContract;
+use MoonShine\Core\Exceptions\ResourceException;
 use MoonShine\Laravel\Collections\Fields;
 use MoonShine\Laravel\DependencyInjection\MoonShine;
 use MoonShine\Laravel\Fields\Relationships\ModelRelationField;
 use MoonShine\Laravel\Resources\ModelResource;
 use MoonShine\Laravel\Traits\Request\HasPageRequest;
 use MoonShine\Laravel\Traits\Request\HasResourceRequest;
+use MoonShine\Support\Enums\Ability;
+use MoonShine\Support\Enums\Action;
 use MoonShine\Support\Enums\PageType;
 use MoonShine\UI\Exceptions\FieldException;
 use Throwable;
@@ -25,6 +28,40 @@ class RelationModelFieldRequest extends FormRequest
     use HasResourceRequest;
     /** @use HasPageRequest<CrudPageContract<ModelResource, MoonShine, Fields>> */
     use HasPageRequest;
+
+    /**
+     * @throws Throwable
+     */
+    public function authorize(): bool
+    {
+        $resource = $this->getResource();
+
+        if (\is_null($resource)) {
+            throw ResourceException::notDeclared();
+        }
+
+        $ability = match ($this->getPage()->getPageType()) {
+            PageType::INDEX => Ability::VIEW_ANY,
+            PageType::DETAIL => Ability::VIEW,
+            PageType::FORM => $this->route('resourceItem') === null
+                ? Ability::CREATE
+                : Ability::UPDATE,
+            default => Ability::VIEW_ANY,
+        };
+
+        $action = match ($ability) {
+            Ability::VIEW => Action::VIEW,
+            Ability::CREATE => Action::CREATE,
+            Ability::UPDATE => Action::UPDATE,
+            default => null,
+        };
+
+        if ($action !== null && ! $resource->hasAction($action)) {
+            return false;
+        }
+
+        return $resource->can($ability);
+    }
 
     public function getRelationName(): string
     {
