+++ tests/Feature/Controllers/HasManyControllerTest.php
@@ -5,12 +5,15 @@ declare(strict_types=1);
 use MoonShine\Contracts\Core\DependencyInjection\CoreContract;
 use MoonShine\Laravel\Fields\Relationships\BelongsToMany;
 use MoonShine\Laravel\Fields\Relationships\HasMany;
+use MoonShine\Laravel\Models\MoonshineUser;
 use MoonShine\Support\Enums\PageType;
 use MoonShine\Tests\Fixtures\Models\Category;
 use MoonShine\Tests\Fixtures\Models\Item;
 use MoonShine\Tests\Fixtures\Resources\TestCategoryResource;
+use MoonShine\Tests\Fixtures\Resources\TestCommentResource;
 use MoonShine\Tests\Fixtures\Resources\TestItemResource;
 use MoonShine\Tests\Fixtures\Resources\TestResource;
+use MoonShine\Tests\Fixtures\Resources\TestResourceBuilder;
 use MoonShine\UI\Fields\ID;
 use MoonShine\UI\Fields\Text;
 
@@ -121,6 +124,29 @@ it('get form component', function () {
         ->assertOk()
         ->assertSee('form')
     ;
+});
+
+it('forbids relation data when update policy denies access to the parent resource', function (): void {
+    $item = createItem(countComments: 1);
+
+    $resource = TestResourceBuilder::new(Item::class)
+        ->setTestFields([
+            ID::make(),
+            HasMany::make('Comments', resource: TestCommentResource::class),
+        ])
+        ->setTestPolicy(true);
+
+    MoonshineUser::query()->whereKey(1)->update([
+        'name' => 'Policies test',
+    ]);
+
+    asAdmin()->get($this->moonshineCore->getRouter()->to('has-many.list', [
+        'pageUri' => PageType::FORM->value,
+        'resourceUri' => $resource->getUriKey(),
+        'resourceItem' => $item->getKey(),
+        '_relation' => 'comments',
+    ]))
+        ->assertForbidden();
 });
 
 it('uses related resource context for nested relation fields in form component', function (): void {
